Move logic for file upload to firebase storage to separate services file
